+++ sonar-plugin/org.wso2.code.quality.sonar.plugin/src/main/java/org/wso2/code/quality/sonar/plugin/RulesPlugin.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2023, WSO2 LLC. (http://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.wso2.code.quality.sonar.plugin;
+
+import org.sonar.api.Plugin;
+
+/**
+ * Entry point of your plugin containing your custom rules.
+ */
+public class RulesPlugin implements Plugin {
+
+    @Override
+    public void define(Context context) {
+
+        // Server extensions -> objects are instantiated during server startup.
+        context.addExtension(RulesDefinitionImpl.class);
+
+        // Batch extensions -> objects are instantiated during code analysis.
+        context.addExtension(FileCheckRegistrar.class);
+    }
+}
