+++ sonar-plugin/org.wso2.code.quality.sonar.plugin/src/main/resources/org/wso2/code/quality/sonar/plugin/rules/descriptions/IncInComments.json
